+++ modbus-mqtt/src/modbus/connector.rs
@@ -2,7 +2,7 @@ use crate::modbus::{connection, register};
 use crate::mqtt::{Payload, Scopable};
 use crate::{mqtt, shutdown::Shutdown};
 use serde::Deserialize;
-use serde_json::value::RawValue as RawJSON;
+use serde_json::value::Value as JSON;
 use tokio::select;
 use tracing::{debug, error, info};
 
@@ -81,34 +81,45 @@ async fn parse_and_connect(
     }
     Ok(())
 }
-async fn connect(config: Config<'_>, mqtt: mqtt::Handle, shutdown: Shutdown) -> crate::Result<()> {
+async fn connect(config: Config, mqtt: mqtt::Handle, shutdown: Shutdown) -> crate::Result<()> {
     if shutdown.is_shutdown() {
         return Ok(());
     }
 
+    #[allow(deprecated)]
     let Config {
         connection: settings,
         input,
         holding,
+        registers,
     } = config;
 
     let _ = connection::run(settings, mqtt.clone(), shutdown).await?;
 
     // TODO: consider waiting 1 second before sending the registers to MQTT, to ensure that the connection is listening.
 
-    for (reg_type, registers) in [("holding", holding), ("input", input)] {
-        let mqtt = mqtt.scoped(reg_type);
+    enum Type {
+        Holding,
+        Input,
+        Unchanged,
+    }
+    for (reg_type, registers) in [
+        (Type::Holding, holding),
+        (Type::Input, input),
+        (Type::Unchanged, registers),
+    ] {
+        use register::*;
+        let mqtt = mqtt.scoped("registers");
         for reg in registers {
-            if let Ok(r) =
-                serde_json::from_slice::<register::AddressedRegister>(reg.get().as_bytes())
-            {
-                let json = serde_json::to_vec(&r.register).unwrap(); // unwrap() should be fine because we JUST deserialized it successfully
-                mqtt.publish(r.address.to_string(), json).await?;
-                // if let Some(name) = r.register.name {
-                //     r.register.name = None;
-                //     let json = serde_json::to_vec(&r).unwrap(); // unwrap() should be fine because we JUST deserialized it successfully
-                //     mqtt.publish(name, json).await?;
-                // }
+            if let Ok(mut reg) = serde_json::from_value::<Register>(reg) {
+                reg.register_type = match reg_type {
+                    Type::Holding => RegisterType::Holding,
+                    Type::Input => RegisterType::Input,
+                    Type::Unchanged => reg.register_type,
+                };
+
+                let json = serde_json::to_vec(&reg).unwrap(); // unwrap() should be fine because we JUST deserialized it successfully
+                mqtt.publish(format!("{}/config", reg.path()), json).await?;
             }
         }
     }
@@ -119,15 +130,22 @@ async fn connect(config: Config<'_>, mqtt: mqtt::Handle, shutdown: Shutdown) -> 
 /// Wrapper around `modbus::connection::Config` that can include some registers inline, which the connector will
 /// re-publish to the appropriate topic once the connection is established.
 #[derive(Debug, Deserialize)]
-struct Config<'a> {
+struct Config {
     #[serde(flatten)]
     connection: connection::Config,
 
     // Allow registers to be defined inline, but capture them as raw JSON so that if they have incorrect schema, we can
     // still establish the Modbus connection. Valid registers will be re-emitted as individual register configs to MQTT,
     // to be picked up by the connection.
-    #[serde(default, borrow)]
-    pub input: Vec<&'a RawJSON>,
-    #[serde(alias = "hold", default, borrow)]
-    pub holding: Vec<&'a RawJSON>,
+    #[deprecated]
+    #[serde(default)]
+    input: Vec<JSON>,
+
+    #[deprecated]
+    #[serde(alias = "hold", default)]
+    holding: Vec<JSON>,
+
+    #[deprecated]
+    #[serde(default)]
+    registers: Vec<JSON>,
 }